+++ stroom-core-shared/src/main/java/stroom/job/shared/ScheduledTimeResource.java
@@ -18,6 +18,7 @@ package stroom.job.shared;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiParam;
 import org.fusesource.restygwt.client.DirectRestService;
 import stroom.util.shared.ResourcePaths;
 import stroom.util.shared.RestResource;
@@ -33,9 +34,10 @@ import javax.ws.rs.core.MediaType;
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
 public interface ScheduledTimeResource extends RestResource, DirectRestService {
-    @POST
+    @
+    POST
     @ApiOperation(
             value = "Gets scheduled time info",
             response = ScheduledTimes.class)
-    ScheduledTimes get(GetScheduledTimesRequest request);
+    ScheduledTimes get(@ApiParam("request") GetScheduledTimesRequest request);
 }