+++ stroom-core-shared/src/main/java/stroom/meta/shared/MetaResource.java
@@ -18,6 +18,7 @@ package stroom.meta.shared;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiParam;
 import org.fusesource.restygwt.client.DirectRestService;
 import stroom.util.shared.ResourceGeneration;
 import stroom.util.shared.ResourcePaths;
@@ -51,7 +52,7 @@ public interface MetaResource extends RestResource, DirectRestService {
     @ApiOperation(
             value = "Find matching meta data",
             response = ResourceGeneration.class)
-    ResultPage<MetaRow> findMetaRow(FindMetaCriteria criteria);
+    ResultPage<MetaRow> findMetaRow(@ApiParam("criteria") FindMetaCriteria criteria);
 
     @GET
     @Path("info/{id}")